Change From<Float> to TryFrom<Float>
Erlang ETF explicitly forbids the use of non-finite floats in
`NEW_FLOAT_EXT` and `FLOAT_EXT`, so this should be caught both on
unpacking as well as on writing. This means, however, that `Float` can
not always be converted from `f32` and `f64`, so the respective `From`
implementations have to be replaced by `TryFrom`.